+++ datafusion/functions-nested/Cargo.toml
@@ -58,8 +58,6 @@ rand = "0.8.5"
 
 [dev-dependencies]
 criterion = { version = "0.5", features = ["async_tokio"] }
-datafusion = { workspace = true }
-tokio = { workspace = true, features = ["macros", "rt", "sync"] }
 
 [[bench]]
 harness = false
